+++ example/lib/interface/knight_interface.dart
@@ -3,7 +3,6 @@ import 'dart:ui';
 import 'package:bonfire/bonfire.dart';
 import 'package:example/interface/bar_life_component.dart';
 import 'package:example/player/knight.dart';
-import 'package:flutter/material.dart';
 
 class KnightInterface extends GameInterface {
   @override